+++ changelog.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## NEXT
+
+- Fixed a bug in the `cai_causal_graph.time_series_causal_graph.TimeSeriesCausalGraph.get_topological_order` method in
+  `cai_causal_graph.time_series_causal_graph.TimeSeriesCausalGraph`.
+
 ## 0.3.8
 
 - Fixed a bug in `cai_causal_graph.time_series_causal_graph.TimeSeriesCausalGraph.from_causal_graph` for